Add support for multiple mime types for existing images sizes.
For every single image size a new mime type would be created
if the original mime type of the image supports the specified
mime type.

Currently this only works with: `image/jpg` and `image/webp`
and vice-versa. Only for all the available image sizes except
the full size image.

This PR fixes: #142, #8