Create set of tests to replicate the reported bug
When an image is edited and restored, the full size sources
are not stored in the meta `_wp_attachment_backup_sources` with
this change we introduce the storage of those values when
the constant IMAGE_EDIT_OVERWRITE is not defined or it contains
a falsy value.

In order to follow the same patter WordPress uses to store the backup
sizes.